app/requirements.txt
fastapi[standard]
httpx
pika
python-dotenv
pydantic
prometheus-fastapi-instrumentator
prometheus-client